chore(Tests): Moved avatar and badge files.
