+++ src/components/avatar/avatar.spec.ts
@@ -1,6 +1,6 @@
 import { html, fixture, expect, elementUpdated } from '@open-wc/testing';
-import { IgcAvatarComponent } from '../src/components/avatar/avatar';
-import '../igniteui-webcomponents.js';
+import { IgcAvatarComponent } from './avatar';
+import '../../../igniteui-webcomponents.js';
 
 describe('Avatar', () => {
   it('passes the a11y audit', async () => {
